refactor(core): rename embaddable_runbook to embedded_runbook for consistency
